+++ apps/router-e2e/__e2e__/02-server-actions/components/server-actions-in-file.tsx
@@ -1,0 +1,3 @@
+'use server';
+
+export const greet = async (inputName: string) => `Hello ${inputName} from server!`;