+++ packages/expo-updates/ios/EXUpdates/Database/Migrations/UpdatesDatabaseMigration5To6.swift
@@ -50,3 +50,5 @@ internal final class UpdatesDatabaseMigration5To6: UpdatesDatabaseMigration {
     }
   }
 }
+
+// swiftlint:enable line_length