+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoManager.kt
@@ -4,6 +4,7 @@ import androidx.annotation.OptIn
 import androidx.media3.common.util.UnstableApi
 import expo.modules.kotlin.AppContext
 import expo.modules.video.player.VideoPlayer
+import java.lang.ref.WeakReference
 
 // Helper class used to keep track of all existing VideoViews and VideoPlayers
 @OptIn(UnstableApi::class)
@@ -12,6 +13,7 @@ object VideoManager {
 
   // Used for sharing videoViews between VideoView and FullscreenPlayerActivity
   private var videoViews = mutableMapOf<String, VideoView>()
+  private var fullscreenPlayerActivities = mutableMapOf<String, WeakReference<FullscreenPlayerActivity>>()
 
   // Keeps track of all existing VideoPlayers, and whether they are attached to a VideoView
   private var videoPlayersToVideoViews = mutableMapOf<VideoPlayer, MutableList<VideoView>>()
@@ -44,6 +46,14 @@ object VideoManager {
     audioFocusManager.unregisterPlayer(videoPlayer)
   }
 
+  fun registerFullscreenPlayerActivity(id: String, fullscreenActivity: FullscreenPlayerActivity) {
+    fullscreenPlayerActivities[id] = WeakReference(fullscreenActivity)
+  }
+
+  fun unregisterFullscreenPlayerActivity(id: String) {
+    fullscreenPlayerActivities.remove(id)
+  }
+
   fun onVideoPlayerAttachedToView(videoPlayer: VideoPlayer, videoView: VideoView) {
     if (videoPlayersToVideoViews[videoPlayer]?.contains(videoView) == true) {
       return
@@ -70,15 +80,40 @@ object VideoManager {
     return videoPlayersToVideoViews[videoPlayer]?.isNotEmpty() ?: false
   }
 
-  fun onAppForegrounded() = Unit
+  fun onAppForegrounded() {
+    for (videoView in videoViews.values) {
+      videoView.playerView.useController = videoView.useNativeControls
+    }
+
+    // Pressing the app icon will bring up the mainActivity instead of the fullscreen activity (at least for BareExpo)
+    // In this case we have to manually finish the fullscreen activity
+    for (fullscreenActivity in fullscreenPlayerActivities.values) {
+      fullscreenActivity.get()?.finish()
+    }
+  }
 
   fun onAppBackgrounded() {
     for (videoView in videoViews.values) {
-      if (videoView.videoPlayer?.staysActiveInBackground == false &&
-        !videoView.willEnterPiP &&
-        !videoView.isInFullscreen
-      ) {
-        videoView.videoPlayer?.player?.pause()
+      if (shouldPauseVideo(videoView)) {
+        handleVideoPause(videoView)
+      } else {
+        videoView.wasAutoPaused = false
+      }
+    }
+  }
+
+  private fun shouldPauseVideo(videoView: VideoView): Boolean {
+    return videoView.videoPlayer?.staysActiveInBackground == false &&
+      !videoView.willEnterPiP &&
+      !videoView.isInFullscreen
+  }
+
+  private fun handleVideoPause(videoView: VideoView) {
+    videoView.playerView.useController = false
+    videoView.videoPlayer?.player?.let { player ->
+      if (player.isPlaying) {
+        player.pause()
+        videoView.wasAutoPaused = true
       }
     }
   }