+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -22,6 +22,7 @@ import expo.modules.video.records.BufferOptions
 import expo.modules.video.records.VideoSource
 import expo.modules.video.utils.ifYogaDefinedUse
 import expo.modules.video.utils.makeYogaUndefinedIfNegative
+import expo.modules.video.utils.runWithPiPMisconfigurationSoftHandling
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.launch
@@ -135,7 +136,7 @@ class VideoModule : Module() {
       }
 
       AsyncFunction("startPictureInPicture") { view: VideoView ->
-        view.runWithPiPMisconfigurationSoftHandling(true) {
+        runWithPiPMisconfigurationSoftHandling(true) {
           view.enterPictureInPicture()
         }
       }