+++ packages/expo/src/Expo.ts
@@ -18,4 +18,4 @@ export {
   reloadAppAsync,
 } from 'expo-modules-core';
 
-export { useEvent } from './hooks/useEvent';
+export { useEvent, useEventListener } from './hooks/useEvent';